improve usability of weight input and prevent negative values
- Refactored weight and age inputs to use `int` instead of `double` for cleaner handling
- Improved UX of the increment button for weight input
- Prevented age and weight from going below 0 by:
  - Disabling and hiding the minus button when values are 0
  - Ensuring state logic enforces non-negative values
- Cleaned up and refactored related logic for better maintainability

+++ lib/widgets/information_widget.dart
@@ -11,18 +11,18 @@ class InformationWidget extends StatefulWidget {
       required this.onChangedWeight,
       required this.onChangedage,
       required this.onChangedGender});
-  final Function(double weight) onChangedWeight;
+  final Function(int weight) onChangedWeight;
   final Function(double height) onChangedHeight;
-  final Function(double age) onChangedage;
+  final Function(int age) onChangedage;
   final Function(String gender) onChangedGender;
   @override
   State<InformationWidget> createState() => _InformationWidgetState();
 }
 
 class _InformationWidgetState extends State<InformationWidget> {
-  double weight = 0;
+  int weight = 0;
   double height = 0;
-  double age = 0;
+  int age = 0;
   String gender = '';
   @override
   Widget build(BuildContext context) {
